Add local network interface for tests
