+++ tests/interfaces/network.go
@@ -3,25 +3,10 @@ package interfaces
 import (
 	"context"
 	"crypto/ecdsa"
-	"math/big"
 
-	"github.com/ava-labs/avalanchego/ids"
 	"github.com/ava-labs/subnet-evm/core/types"
-	"github.com/ava-labs/subnet-evm/ethclient"
-	teleportermessenger "github.com/ava-labs/teleporter/abi-bindings/go/Teleporter/TeleporterMessenger"
 	"github.com/ethereum/go-ethereum/common"
 )
-
-type SubnetTestInfo struct {
-	SubnetID                  ids.ID
-	BlockchainID              ids.ID
-	NodeURIs                  []string
-	WSClient                  ethclient.Client
-	RPCClient                 ethclient.Client
-	EVMChainID                *big.Int
-	TeleporterRegistryAddress common.Address
-	TeleporterMessenger       *teleportermessenger.TeleporterMessenger
-}
 
 // Defines the interface for the network setup functions used in the E2E tests
 type Network interface {
