+++ tests/interfaces/subnet_test_info.go
@@ -1,0 +1,22 @@
+package interfaces
+
+import (
+	"math/big"
+
+	"github.com/ava-labs/avalanchego/ids"
+	"github.com/ava-labs/subnet-evm/ethclient"
+	teleportermessenger "github.com/ava-labs/teleporter/abi-bindings/go/Teleporter/TeleporterMessenger"
+	"github.com/ethereum/go-ethereum/common"
+)
+
+// Tracks information about a test subnet used for executing tests against.
+type SubnetTestInfo struct {
+	SubnetID                  ids.ID
+	BlockchainID              ids.ID
+	NodeURIs                  []string
+	WSClient                  ethclient.Client
+	RPCClient                 ethclient.Client
+	EVMChainID                *big.Int
+	TeleporterRegistryAddress common.Address
+	TeleporterMessenger       *teleportermessenger.TeleporterMessenger
+}
